refactor: use "lib" with package imports
