+++ bricks/dart_frog_prod_server/hooks/analysis_options.yaml
@@ -1,1 +1,4 @@
 include: package:very_good_analysis/analysis_options.5.1.0.yaml
+linter:
+  rules:
+    public_member_api_docs: false
